added KeyManager class
